Actualizacion general de componentes, servicios y vistas del frontend 18:05

+++ frontend/src/services/friendshipService.js
@@ -10,14 +10,18 @@ const collectionRef = collection(db, collectionString)
 // GETALL
 export const getAll = async ()=>{
     const data = await getDocs(collectionRef)
+
     return data
 }
 
 //GETBYID
 export const getById = async id =>{
     const document = await getDoc(doc(db,collectionString,id))
-    if (document.exists()) return document
+
+    if (document.exists) return document
+
     return null
+
 }
 
 //CREATE
@@ -26,17 +30,21 @@ export const createFriendship = async document => await addDoc(collectionRef,doc
 //UPDATE
 export const updateFriendship = async (id, documento) => {
     const docRef = doc(db,collectionString,id)
-    return await updateDoc(docRef,documento)
+
+    if (docRef.exists) 
+        return await updateDoc(docRef,documento)
 }
 
 //DELETE
 export const deleteFriendship = async id => {
     const docRef = doc(db,collectionString,id)
-    return await deleteDoc(docRef)
+
+    if (docRef.exists) 
+        return await deleteDoc(docRef)
 }
 
 // ============================================================
-// FUNCIONES ESPECÍFICAS DE SOLICITUDES DE AMISTAD (ÓRBITAS)
+// FUNCIONES ESPECÍFICAS DE SOLICITUDES DE AMISTAD
 // ============================================================
 
 // Enviar una solicitud de amistad de "solicitanteId" hacia "receptorId"
